+++ MiniApp/Classes/core/Extensions/URL+MiniApp.swift
@@ -19,4 +19,20 @@ extension URL {
                 .joined(separator: ".")
                 .trimmingCharacters(in: CharacterSet(charactersIn: "."))
     }
+
+    var isBase64: Bool {
+        let text = self.absoluteString
+        let capturePattern = #"^data:([a-z]+)\/([a-z]+);([\S]*;)?base64,"#
+        guard
+            let captureRegex = try? NSRegularExpression(
+                pattern: capturePattern,
+                options: []
+            )
+        else { return false }
+        let textRange = NSRange(
+            text.startIndex..<text.endIndex,
+            in: text
+        )
+        return captureRegex.firstMatch(in: text, options: [], range: textRange) != nil
+    }
 }